Extracted majority of install command internals to Processor class
